Refactor of HttpMessageHandler based instrumentation.
