--- src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Http/CurlHandler/HttpMethodStruct.cs
@@ -1,16 +1,0 @@
-using Datadog.Trace.DuckTyping;
-
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.CurlHandler
-{
-    /// <summary>
-    /// Http method struct copy target for ducktyping
-    /// </summary>
-    [DuckCopy]
-    public struct HttpMethodStruct
-    {
-        /// <summary>
-        /// Gets the http method in string
-        /// </summary>
-        public string Method;
-    }
-}
