+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Http/HttpClient/CurlHandler/CurlHandlerIntegration.cs
@@ -3,23 +3,26 @@ using System.Threading;
 using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Configuration;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.SocketsHttpHandler
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.HttpClient.CurlHandler
 {
     /// <summary>
-    /// System.Net.Http.SocketsHttpHandler calltarget instrumentation
+    /// System.Net.Http.CurlHandler calltarget instrumentation
     /// </summary>
     [InstrumentMethod(
         AssemblyName = "System.Net.Http",
-        TypeName = "System.Net.Http.SocketsHttpHandler",
-        MethodName = "Send",
-        ReturnTypeName = ClrNames.HttpResponseMessage,
+        TypeName = "System.Net.Http.CurlHandler",
+        MethodName = "SendAsync",
+        ReturnTypeName = ClrNames.HttpResponseMessageTask,
         ParameterTypeNames = new[] { ClrNames.HttpRequestMessage, ClrNames.CancellationToken },
-        MinimumVersion = "5.0.0",
-        MaximumVersion = "5.*.*",
+        MinimumVersion = "4.0.0",
+        MaximumVersion = "4.*.*",
         IntegrationName = IntegrationName)]
-    public class SocketsHttpHandlerSyncIntegration
+    public class CurlHandlerIntegration
     {
         private const string IntegrationName = nameof(IntegrationIds.HttpMessageHandler);
+        private static readonly IntegrationInfo IntegrationId = IntegrationRegistry.GetIntegrationInfo(IntegrationName);
+        private static readonly IntegrationInfo CurlHandlerIntegrationId = IntegrationRegistry.GetIntegrationInfo(nameof(IntegrationIds.CurlHandler));
+        private static readonly Func<bool> IsIntegrationEnabledFunc = () => Tracer.Instance.Settings.IsIntegrationEnabled(CurlHandlerIntegrationId, defaultValue: false);
 
         /// <summary>
         /// OnMethodBegin callback
@@ -33,7 +36,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.SocketsHttpHandler
         public static CallTargetState OnMethodBegin<TTarget, TRequest>(TTarget instance, TRequest requestMessage, CancellationToken cancellationToken)
             where TRequest : IHttpRequestMessage
         {
-            return SocketsHttpHandlerCommon.OnMethodBegin(instance, requestMessage, cancellationToken);
+            return HttpMessageHandlerCommon.OnMethodBegin(instance, requestMessage, cancellationToken, IntegrationId, IsIntegrationEnabledFunc);
         }
 
         /// <summary>
@@ -46,11 +49,10 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.SocketsHttpHandler
         /// <param name="exception">Exception instance in case the original code threw an exception.</param>
         /// <param name="state">Calltarget state value</param>
         /// <returns>A response value, in an async scenario will be T of Task of T</returns>
-        public static CallTargetReturn<TResponse> OnMethodEnd<TTarget, TResponse>(TTarget instance, TResponse responseMessage, Exception exception, CallTargetState state)
+        public static TResponse OnAsyncMethodEnd<TTarget, TResponse>(TTarget instance, TResponse responseMessage, Exception exception, CallTargetState state)
             where TResponse : IHttpResponseMessage
         {
-            var response = SocketsHttpHandlerCommon.OnMethodEnd(instance, responseMessage, exception, state);
-            return new CallTargetReturn<TResponse>(response);
+            return HttpMessageHandlerCommon.OnMethodEnd(instance, responseMessage, exception, state);
         }
     }
 }
