+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Http/HttpClient/IHttpResponseMessage.cs
@@ -1,9 +1,11 @@
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WinHttpHandler
+using Datadog.Trace.DuckTyping;
+
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.HttpClient
 {
     /// <summary>
     /// HttpResponseMessage interface for ducktyping
     /// </summary>
-    public interface IHttpResponseMessage
+    public interface IHttpResponseMessage : IDuckType
     {
         /// <summary>
         /// Gets the status code of the http response
