+++ src/Datadog.Trace.ClrProfiler.Managed/AutoInstrumentation/Http/HttpClient/WinHttpHandler/WinHttpHandlerIntegration.cs
@@ -1,12 +1,9 @@
 using System;
-using System.Linq;
 using System.Threading;
 using Datadog.Trace.ClrProfiler.CallTarget;
 using Datadog.Trace.Configuration;
-using Datadog.Trace.ExtensionMethods;
-using Datadog.Trace.Tagging;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WinHttpHandler
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.HttpClient.WinHttpHandler
 {
     /// <summary>
     /// System.Net.Http.WinHttpHandler calltarget instrumentation
@@ -25,6 +22,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WinHttpHandler
         private const string IntegrationName = nameof(IntegrationIds.HttpMessageHandler);
         private static readonly IntegrationInfo IntegrationId = IntegrationRegistry.GetIntegrationInfo(IntegrationName);
         private static readonly IntegrationInfo WinHttpHandlerIntegrationId = IntegrationRegistry.GetIntegrationInfo(nameof(IntegrationIds.WinHttpHandler));
+        private static readonly Func<bool> IsIntegrationEnabledFunc = () => Tracer.Instance.Settings.IsIntegrationEnabled(WinHttpHandlerIntegrationId, defaultValue: false);
 
         /// <summary>
         /// OnMethodBegin callback
@@ -38,21 +36,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WinHttpHandler
         public static CallTargetState OnMethodBegin<TTarget, TRequest>(TTarget instance, TRequest requestMessage, CancellationToken cancellationToken)
             where TRequest : IHttpRequestMessage
         {
-            if (IsTracingEnabled(requestMessage.Headers))
-            {
-                Scope scope = ScopeFactory.CreateOutboundHttpScope(Tracer.Instance, requestMessage.Method.Method, requestMessage.RequestUri, IntegrationId, out HttpTags tags);
-                if (scope != null)
-                {
-                    tags.HttpClientHandlerType = instance.GetType().FullName;
-
-                    // add distributed tracing headers to the HTTP request
-                    SpanContextPropagator.Instance.Inject(scope.Span.Context, new HttpHeadersCollection(requestMessage.Headers));
-
-                    return new CallTargetState(scope);
-                }
-            }
-
-            return CallTargetState.GetDefault();
+            return HttpMessageHandlerCommon.OnMethodBegin(instance, requestMessage, cancellationToken, IntegrationId, IsIntegrationEnabledFunc);
         }
 
         /// <summary>
@@ -68,60 +52,7 @@ namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WinHttpHandler
         public static TResponse OnAsyncMethodEnd<TTarget, TResponse>(TTarget instance, TResponse responseMessage, Exception exception, CallTargetState state)
             where TResponse : IHttpResponseMessage
         {
-            Scope scope = state.Scope;
-
-            if (scope is null)
-            {
-                return responseMessage;
-            }
-
-            try
-            {
-                scope.Span.SetHttpStatusCode(responseMessage.StatusCode, isServer: false);
-
-                if (exception != null)
-                {
-                    scope.Span.SetException(exception);
-                }
-            }
-            finally
-            {
-                scope.Dispose();
-            }
-
-            return responseMessage;
-        }
-
-        private static bool IsTracingEnabled(IRequestHeaders headers)
-        {
-            if (!Tracer.Instance.Settings.IsIntegrationEnabled(WinHttpHandlerIntegrationId, defaultValue: false))
-            {
-                return false;
-            }
-
-            if (headers.TryGetValues(HttpHeaderNames.TracingEnabled, out var headerValues))
-            {
-                if (headerValues is string[] arrayValues)
-                {
-                    for (var i = 0; i < arrayValues.Length; i++)
-                    {
-                        if (string.Equals(arrayValues[i], "false", StringComparison.OrdinalIgnoreCase))
-                        {
-                            return false;
-                        }
-                    }
-
-                    return true;
-                }
-
-                if (headerValues != null && headerValues.Any(s => string.Equals(s, "false", StringComparison.OrdinalIgnoreCase)))
-                {
-                    // tracing is disabled for this request via http header
-                    return false;
-                }
-            }
-
-            return true;
+            return HttpMessageHandlerCommon.OnMethodEnd(instance, responseMessage, exception, state);
         }
     }
 }
